requirements.txt
sentence-transformers
langchain
langchain-community
langchain-core
flask
pypdf==3.17.1
python-dotenv
pinecone
langchain-pinecone
langchain_openai
langchain_experimental
openai==1.3.0
huggingface-hub
torch>=2.0.0
transformers
langsmith
Pillow>=10.0.0

-e .


